requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# Database and ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI related
openai==1.3.7

# Enterprise Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
PyJWT==2.8.0

# Logging and monitoring
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-requests==0.57b0
prometheus-client==0.19.0
psutil==5.9.6

# Rate limiting and security
slowapi==0.1.9
redis==5.0.1

# Enterprise utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
httpx==0.25.2
cachetools==5.3.2
python-json-logger==2.0.7

# Enterprise monitoring and alerting
watchdog==3.0.0
structlog==23.2.0

# Performance and optimization
uvloop==0.19.0
httptools==0.6.1

# Testing and quality (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1